requirements.txt
aiogram
python-dotenv
feedparser
requests
aiocron
pytest